profileservice: add flow with aac

+++ sco.small.profileservice/src/main/java/it/smartcommunitylab/small/profileservice/controller/rest/ExtendedProfileController.java
@@ -1,6 +1,7 @@
 package it.smartcommunitylab.small.profileservice.controller.rest;
 
 import io.swagger.annotations.ApiOperation;
+import io.swagger.annotations.Authorization;
 import it.smartcommunitylab.small.profileservice.common.EntityNotFoundException;
 import it.smartcommunitylab.small.profileservice.common.UnauthorizedException;
 import it.smartcommunitylab.small.profileservice.common.Utils;
@@ -9,12 +10,14 @@ import it.smartcommunitylab.small.profileservice.model.ExtendedProfile;
 
 import java.util.Map;
 
+import javax.annotation.PostConstruct;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.ExceptionHandler;
@@ -25,12 +28,25 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
+import eu.trentorise.smartcampus.aac.AACService;
+
 @Controller
 public class ExtendedProfileController {
 	private static final transient Logger logger = LoggerFactory.getLogger(ExtendedProfileController.class);
 	
 	@Autowired
+	@Value("${ext.aacURL}")
+	private String aacUrl;
+	
+	@Autowired
 	private RepositoryManager repository;
+	
+	private AACService aacService;
+	
+	@PostConstruct
+	public void init() {
+		aacService = new AACService(aacUrl, null, null);
+	}
 	
 //	@RequestMapping(method = RequestMethod.GET, value = "/extprofile/me")
 //	@ApiOperation(value = "get logged user's profiles", response=ExtendedProfile.class, responseContainer="List")
@@ -48,11 +64,17 @@ public class ExtendedProfileController {
 //	}
 	
 	@RequestMapping(method = RequestMethod.GET, value = "/extprofile/me/{profileId}")
-	@ApiOperation(value = "get a spacific logged user's profile")
+	@ApiOperation(value = "get a spacific logged user's profile", 
+		authorizations = {@Authorization(value = "Authorization")})
 	public @ResponseBody ExtendedProfile getMyProfileById(@PathVariable String profileId, HttpServletRequest request,
 			HttpServletResponse response) throws Exception {
-//		String userId = Utils.getUserId();
-		String userId = "test";
+		String userId = Utils.getUserId();
+		if(Utils.isNull(userId)) {
+			throw new UnauthorizedException("token not valid or requested scope non present");
+		}
+		if(!Utils.isRequestScopePermitted(request, aacService, profileId, "personal", "read")) {
+			throw new UnauthorizedException("token not valid or requested scope non present");
+		}
 		if(logger.isInfoEnabled()) {
 			logger.info(String.format("getMyProfileById - %s - %s", userId, profileId));
 		}
@@ -61,49 +83,58 @@ public class ExtendedProfileController {
 	}
 	
 	@RequestMapping(method = RequestMethod.POST, value = "/extprofile/me/{profileId}")
-	@ApiOperation(value = "add a new profile to logged user")
+	@ApiOperation(value = "add a new profile to logged user",
+		authorizations = {@Authorization(value = "Authorization")})
 	public @ResponseBody ExtendedProfile addMyProfile(@PathVariable String profileId,	@RequestBody ExtendedProfile profile, 
 			HttpServletRequest request, HttpServletResponse response) throws Exception {
-//		String userId = Utils.getUserId();
-		String userId = "test";
+		String userId = Utils.getUserId();
+		if(Utils.isNull(userId)) {
+			throw new UnauthorizedException("token not valid or requested scope non present");
+		}
+		if(!Utils.isRequestScopePermitted(request, aacService, profileId, "personal", "write")) {
+			throw new UnauthorizedException("token not valid or requested scope non present");
+		}
 		if(logger.isInfoEnabled()) {
 			logger.info(String.format("addMyProfile - %s - %s", userId, profileId));
 		}
-//		if(!Utils.isRequestScopePermitted(request, profileId, "personal", "write")) {
-//			throw new UnauthorizedException("requested scope non present");
-//		}
 		ExtendedProfile result = repository.saveExtendedProfile(userId, profileId, profile);
 		return result;
 	}
 
 	@RequestMapping(method = RequestMethod.PUT, value = "/extprofile/me/{profileId}")
-	@ApiOperation(value = "update a spacific logged user's profile")
+	@ApiOperation(value = "update a spacific logged user's profile",
+		authorizations = {@Authorization(value = "Authorization")})
 	public @ResponseBody ExtendedProfile updateMyProfile(@PathVariable String profileId, @RequestBody ExtendedProfile profile, 
 			HttpServletRequest request, HttpServletResponse response) throws Exception {
-//		String userId = Utils.getUserId();
-		String userId = "test";
+		String userId = Utils.getUserId();
+		if(Utils.isNull(userId)) {
+			throw new UnauthorizedException("token not valid or requested scope non present");
+		}
+		if(!Utils.isRequestScopePermitted(request, aacService, profileId, "personal", "write")) {
+			throw new UnauthorizedException("token not valid or requested scope non present");
+		}
 		if(logger.isInfoEnabled()) {
 			logger.info(String.format("updateMyProfile - %s - %s", userId, profileId));
 		}
-//		if(!Utils.isRequestScopePermitted(request, profileId, "personal", "write")) {
-//			throw new UnauthorizedException("requested scope non present");
-//		}
 		ExtendedProfile result = repository.updateExtendedProfile(userId, profileId, profile);
 		return result;
 	}
 
 	@RequestMapping(method = RequestMethod.DELETE, value = "/extprofile/me/{profileId}")
-	@ApiOperation(value = "delete a spacific logged user's profile")
+	@ApiOperation(value = "delete a spacific logged user's profile",
+		authorizations = {@Authorization(value = "Authorization")})
 	public @ResponseBody ExtendedProfile deleteMyProfile(@PathVariable String profileId, 
 			HttpServletRequest request, HttpServletResponse response) throws Exception {
-//		String userId = Utils.getUserId();
-		String userId = "test";
+		String userId = Utils.getUserId();
+		if(Utils.isNull(userId)) {
+			throw new UnauthorizedException("token not valid or requested scope non present");
+		}
+		if(!Utils.isRequestScopePermitted(request, aacService, profileId, "personal", "write")) {
+			throw new UnauthorizedException("token not valid or requested scope non present");
+		}
 		if(logger.isInfoEnabled()) {
 			logger.info(String.format("deleteMyProfile - %s - %s", userId, profileId));
 		}
-//		if(!Utils.isRequestScopePermitted(request, profileId, "personal", "write")) {
-//			throw new UnauthorizedException("requested scope non present");
-//		}
 		ExtendedProfile profile = repository.deleteExtendedProfile(userId, profileId);
 		return profile;
 	}
@@ -123,9 +154,13 @@ public class ExtendedProfileController {
 //	}
 	
 	@RequestMapping(method = RequestMethod.GET, value = "/extprofile/app/{userId}/{profileId}")
-	@ApiOperation(value = "get a spacific user profile")
+	@ApiOperation(value = "get a spacific user profile",
+		authorizations = {@Authorization(value = "Authorization")})
 	public @ResponseBody ExtendedProfile getUserProfileById(@PathVariable String userId, @PathVariable String profileId, 
 			HttpServletRequest request, HttpServletResponse response) throws Exception {
+		if(!Utils.isRequestScopePermitted(request, aacService, profileId, "app", "read")) {
+			throw new UnauthorizedException("token not valid or requested scope non present");
+		}
 		if(logger.isInfoEnabled()) {
 			logger.info(String.format("getUserProfileById - %s - %s", userId, profileId));
 		}
@@ -134,9 +169,13 @@ public class ExtendedProfileController {
 	}
 	
 	@RequestMapping(method = RequestMethod.POST, value = "/extprofile/app/{userId}/{profileId}")
-	@ApiOperation(value = "add a new user profile")
+	@ApiOperation(value = "add a new user profile",
+		authorizations = {@Authorization(value = "Authorization")})
 	public @ResponseBody ExtendedProfile addUserProfile(@PathVariable String userId, @PathVariable String profileId, 
 			@RequestBody ExtendedProfile profile, HttpServletRequest request, HttpServletResponse response) throws Exception {
+		if(!Utils.isRequestScopePermitted(request, aacService, profileId, "app", "write")) {
+			throw new UnauthorizedException("token not valid or requested scope non present");
+		}
 		if(logger.isInfoEnabled()) {
 			logger.info(String.format("addUserProfile - %s - %s", userId, profileId));
 		}
@@ -145,9 +184,13 @@ public class ExtendedProfileController {
 	}
 		
 	@RequestMapping(method = RequestMethod.PUT, value = "/extprofile/app/{userId}/{profileId}")
-	@ApiOperation(value = "update a spacific user profile")
+	@ApiOperation(value = "update a spacific user profile",
+		authorizations = {@Authorization(value = "Authorization")})
 	public @ResponseBody ExtendedProfile updateUserProfile(@PathVariable String userId, @PathVariable String profileId, 
 			@RequestBody ExtendedProfile profile, HttpServletRequest request, HttpServletResponse response) throws Exception {
+		if(!Utils.isRequestScopePermitted(request, aacService, profileId, "app", "write")) {
+			throw new UnauthorizedException("token not valid or requested scope non present");
+		}
 		if(logger.isInfoEnabled()) {
 			logger.info(String.format("updateUserProfile - %s - %s", userId, profileId));
 		}
@@ -156,9 +199,13 @@ public class ExtendedProfileController {
 	}
 	
 	@RequestMapping(method = RequestMethod.DELETE, value = "/extprofile/app/{userId}/{profileId}")
-	@ApiOperation(value = "delete a spacific user profile")
+	@ApiOperation(value = "delete a spacific user profile",
+		authorizations = {@Authorization(value = "Authorization")})
 	public @ResponseBody ExtendedProfile deleteUserProfile(@PathVariable String userId, @PathVariable String profileId, 
 			HttpServletRequest request, HttpServletResponse response) throws Exception {
+		if(!Utils.isRequestScopePermitted(request, aacService, profileId, "app", "write")) {
+			throw new UnauthorizedException("token not valid or requested scope non present");
+		}
 		if(logger.isInfoEnabled()) {
 			logger.info(String.format("deleteUserProfile - %s - %s", userId, profileId));
 		}
